+++ ast/asdl_rs.py
@@ -288,7 +288,7 @@ class StructVisitor(EmitVisitor):
     def simple_sum(self, sum, name, depth):
         rust_name = rust_type_name(name)
         self.emit_attrs(depth)
-        self.emit("#[derive(is_macro::Is, Copy)]", depth)
+        self.emit("#[derive(is_macro::Is, Copy, Hash, Eq)]", depth)
         self.emit(f"pub enum {rust_name} {{", depth)
         for variant in sum.types:
             self.emit(f"{variant.name},", depth + 1)
@@ -1067,7 +1067,6 @@ def write_visitor_def(mod, type_info, f):
 
 
 def write_ranged_def(mod, type_info, f):
-    f.write("use crate::Ranged;")
     RangedDefVisitor(f, type_info).visit(mod)
