+++ ast/src/gen/ranged.rs
@@ -1,6 +1,5 @@
 // File automatically generated by ast/asdl_rs.py.
 
-use crate::Ranged;
 #[cfg(feature = "all-nodes-with-ranges")]
 impl Ranged for crate::generic::ModModule<TextRange> {
     fn range(&self) -> TextRange {
